Fixed a null-terminator bug
When the file being loaded is smaller in size than the current block size, the editor fills the unused block data with null-terminators, which are then included as if they were a part of the file. This was fixed by resizing the data to the size of the file, rather than the size of a block.