Added simple file saving and improved error handling
I added a very simple and limited way of doing file saving for now. This
saving system is quick but can only overwrite existing data, not
remove. This means that any removed characters will not be removed from
the file on disk, only in memory.

I also improved error handling in worker threads. if an error occurs on
a worker thread, the main thread will be notified, and an error message
will display on the screen.