Implemented tracking of unsaved changes
The title bar now shows a star symbol (*) when there are unsaved changes. If the user tries to exit the application or load a new block, a confirmation will appear to remind the user of their unsaved changes. The new exit dialog will also only show if there are unsaved changes, so if there aren't any changes, the application will close instantly.