Removed 'REPORT_ERROR' macro in workerbase.h
Reverted a previous change to add it as I didn't see a need for it at
the moment. If I need to see where an error occurred in the future,
it's better to create a macro that can be used anywhere in the code,
instead of having one specific only to the WorkerBase class.