Added the ability to open multiple files
It is now possible to select and open multiple files in the open file dialog. The first selected file will open in the same session, and the other files will open in new LFEditor sessions.

This is far from perfect, but it works for now. I will most likely replace this with a tab system where each file gets its own editing tab.